wget_better fix some oversights.
NOTE: Also wget doesnt handle the case, if the local file is larger
than the remote file. So if there is modification ot nav data file,
where the local file is larger than remote file, then the program on
its own wont fetch such file, user will have to manually delete
the file and refetch the file.

+++ mfutilskvc.py
@@ -188,14 +188,24 @@ def proc_days(start, end, handle_date_func, bNotBeyondYesterday=True):
 
 
 def wget_better(url, localFName):
+    """
+    If the file on the server is bigger than the local file,
+    then redownload the file freshly, rather than appending to it,
+    as chances are the local file was not a partial download, but
+    rather a older version of the file with different data.
+    """
     #cmd = "curl {} --remote-time --time-cond {} --output {}".format(url,fName,fName)
-    mtimePrev = os.stat(localFName).st_mtime
+    if os.path.exists(localFName):
+        mtimePrev = os.stat(localFName).st_mtime
+    else:
+        mtimePrev = -1
     cmd = "wget {} --continue --output-document={}".format(url,localFName)
     os.system(cmd)
-    mtimeNow = os.stat(localFName).st_mtime
-    if mtimeNow != mtimePrev:
-        os.remove(localFName)
-        os.system(cmd)
+    if os.path.exists(localFName):
+        mtimeNow = os.stat(localFName).st_mtime
+        if (mtimePrev != -1) and (mtimeNow != mtimePrev):
+            os.remove(localFName)
+            os.system(cmd)
 
 
 def fetch4date(y, m, d):
